Cleanup App component
