+++ front/tsconfig.json
@@ -19,7 +19,8 @@
         "@/*": [
           "./src/*"
         ]
-      }
+      },
+      "types": ["vitest/globals"]
   },
   "include": ["src"],
   "references": [{ "path": "./tsconfig.node.json" }]
